+++ testmod-common/src/main/java/io/github/jamalam360/testmod/TestMod.java
@@ -3,6 +3,9 @@ package io.github.jamalam360.testmod;
 import io.github.jamalam360.jamlib.JamLibPlatform;
 import io.github.jamalam360.jamlib.config.ConfigManager;
 import io.github.jamalam360.jamlib.events.client.ClientPlayLifecycleEvents;
+import io.github.jamalam360.testmod.config.NestedConfigChild;
+import io.github.jamalam360.testmod.config.QuickerConnectButtonTestConfig;
+import io.github.jamalam360.testmod.config.TestConfig;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -14,7 +17,7 @@ public class TestMod {
     public static final ConfigManager<TestConfig> CONFIG_MANAGER = new ConfigManager<>(MOD_ID, "first_config", TestConfig.class);
     public static final ConfigManager<TestConfig> CONFIG_MANAGER_2 = new ConfigManager<>(MOD_ID, "second_config", TestConfig.class);
     public static final ConfigManager<QuickerConnectButtonTestConfig> QCB_CONFIG = new ConfigManager<>(MOD_ID, "quickerconnectbutton", QuickerConnectButtonTestConfig.class);
-
+    public static final ConfigManager<NestedConfigChild> NESTED_CONFIG = new ConfigManager<>(MOD_ID, "nested", NestedConfigChild.class);
 
     public static void init() {
         LOGGER.info("Initializing JamLib Test Mod on {}", JamLibPlatform.getPlatform());