+++ testmod-common/src/main/java/io/github/jamalam360/testmod/config/NestedConfigChild.java
@@ -1,0 +1,5 @@
+package io.github.jamalam360.testmod.config;
+
+public class NestedConfigChild extends NestedConfigParent {
+	public String fromChild = "child";
+}